requirements.txt
aiohttp==3.7.4.post0
asyncpg==0.23.0
httpx==0.16.1
many_stop_words==0.2.2
UpSetPlot==0.6.0
matplotlib==3.4.1
isodate==0.6.0
forex_python==1.5
aiohttp_requests==0.1.3
mecabwrap==0.3.2
seaborn==0.11.1
wordcloud==1.8.1
pandas==1.2.2
youtube_data_api==0.0.20
six==1.16.0
numpy==1.20.3
psycopg2==2.8.6
Pillow==8.3.2
PyQt5==5.15.4
pytz==2021.1
git+https://github.com/KaitoCross/pytchat.git
